Makes the API produce a reference assembly. (#2647)
* Makes the API produce a reference assembly.

A reference assembly is one that has all implementation and dependencies stripped out.  This is very useful for plugin authors because it means they can just reference this one file in their plugin and get fully IDE support, auto-complete, comment documentation, etc. from a single reference that is around 24KB big at the moment.

`Deterministic` should *probably* be applied to this entire repository (helps with auditing final build artifacts), but specifically it is recommended for ReferenceAssemblies which is why I have enabled it here.

Upon building after this change `src\Nethermind\Nethermind.Api\bin\Release\net5.0\ref` will contain a `Nethermind.Api.dll` that is 24KB large.  Users can then reference it in their Plugin projects with:
```xml
<Project Sdk="Microsoft.NET.Sdk">
	<PropertyGroup>
		<TargetFramework>net5.0</TargetFramework>
		<Nullable>enable</Nullable>
	</PropertyGroup>
	<ItemGroup>
		<Reference Include="Nethermind.Api">
			<HintPath>.\Nethermind.Api.dll</HintPath>
			<Private>false</Private>
		</Reference>
	</ItemGroup>
</Project>
```

* Adds all other direct references of PluginAPI.

Did this for myself, figure I would share with the group in case it is desired.