Moves WebSocketsManager instantiation to API Builder.
This makes it so Api.WebSocketsManager can be non-null, which makes it easier to use in plugins who want to use it.  Previously, this was unconditionally instantiated and assigned immediately after the builder finished and there aren't any code paths that set it to null, so it was *effectively* not null but in a way the static analyzer couldn't follow.