Swaps in ConcurrentLru for LruCache.
The current LruCache uses locks for thread synchronization, which can be quite slow.  This change makes it so we use a concurrent LRU instead.

I chose to just wrap the existing LruCache to minimize the changes in this PR.  In a future PR, it would be reasonable to just use `ConcurrentLru` directly, though that would require adding it as a dependency to a lot of projects.

Note: I don't know Nethermind's view on third party libraries.  The referenced library is open source and MIT so we could just copy it over rather than referencing it via NuGet, but that is marginally more work and there is a good chance this change doesn't get accepted so I chose to do less work initially.

Note: The LruCache `name` constructor parameter was never used, so I have removed it as part of this PR since I was touching all of the constructor calls anyway due to removal of the `startCapacity`.