Adds composite tracer factory to blockchain processor. (#2991)
* Adds composite tracer factory to blockchain processor.

This makes it so a plugin can add a block tracer to all new blocks that get processed.  I added it directly to the BlockchainProcessor for two reasons:
1. Adding a new constructor parameter to BlockchainProcessor requires a *ton* wiring.  I can do this if it is desired, but I wanted to start with the simplest solution and get feedback before putting in all of that legwork.
2. `CompositeBlockTracerFactory` is a very thin class (just a reference and a bool) and isn't interesting enough IMO to be considered a generalized service.  In fact, it *may* make sense to just integrate it directly into BlockchainProcessor.

`EndBlockTrace` was added to `IBlockTracer` so that the plugin can correctly aggregate transaction traces by block.  The block is passed as a parameter to this callback just to make it easier on the plugins to correlate start and end of the block trace.

Usage:
```cs
api.BlockchainProcessor.BlockTracerFactory.AddChildFactory(myBlockTracer);
```
This will cause `myBlockTracer` to be attachede to every new block that is processed by the primary block processing pipeline.  In the case where a plugin wants to trace every block (e.g., for building a block explorer), this avoids having to unnecessarily execute the block multiple times (once without tracers, and then again with tracers).

No tests included here, but I can add them if this change otherwise seems reasonable to the maintainers.

* Fixed issues, added two tests

* fixed OneTimeChainProcessor

* fixed BundleBlockTracer

* update tests

* refactorings

* fix tests

Co-authored-by: kristofgazso <kristof.gazso@gmail.com>
Co-authored-by: lukasz.rozmej <lukasz.rozmej@gmail.com>