Adds Deterministic and ProduceReferenceAssembly to Nethermind.Mev (#5438)
All projects should have this set, but this is the one that broke my plugin so it is the one I am submitting a PR for.  Ideally, anytime a new project is created this is automatically added to it.  Alternatively, perhaps there could be a linter check or something that verifies these lines get added to all projects?

It is necessary for anyone building plugins that live external to this repository/project as the reference assemblies are needed to build against.